+++ app.js
@@ -3,7 +3,7 @@ require('dotenv').config({path: process.env.NODE_ENV ? './.env' + process.env.NO
 
 // +++ Adapters +
 const Reddit = require('./src/adapters/reddit')
-const Slack = require('./src/adapters/slack/adapter')
+const Slack = require('./src/adapters/slack/index')
 
 async function bootstrap () {
