+++ src/adapters/slack/index.js
